Allow ttl=0s, so that no update checks are performed.
This falls out automatically from the use of POSIX timers, by
setting the interval to 0.

Whether or not the client nevertheless sees changes in the file
is platform-dependent, since mmap(2) leaves this unspecified for
MAP_PRIVATE. On Linux, changes are seen, and the mapping is
retained after the file is deleted (until munmap).